+++ src/skins/vector/skindex.js
@@ -23,16 +23,19 @@ limitations under the License.
 
 var skin = {};
 
-// TODO: Fix this so matrix-react-sdk stuff is in react SDK.
+// Vector-specific stuff
+skin['elements.Spinner'] = require('../../components/views/elements/Spinner');
+
+// TODO: Fix this so matrix-react-sdk stuff is in react SDK skindex
 skin['avatars.RoomAvatar'] = require('matrix-react-sdk/lib/components/views/avatars/RoomAvatar');
 skin['avatars.MemberAvatar'] = require('matrix-react-sdk/lib/components/views/avatars/MemberAvatar');
 
+// Old style stuff
 skin['atoms.EditableText'] = require('./views/atoms/EditableText');
 skin['atoms.EnableNotificationsButton'] = require('./views/atoms/EnableNotificationsButton');
 skin['atoms.ImageView'] = require('./views/atoms/ImageView');
 skin['atoms.LogoutButton'] = require('./views/atoms/LogoutButton');
 skin['atoms.MessageTimestamp'] = require('./views/atoms/MessageTimestamp');
-skin['atoms.Spinner'] = require('./views/atoms/Spinner');
 skin['atoms.create_room.CreateRoomButton'] = require('./views/atoms/create_room/CreateRoomButton');
 skin['atoms.create_room.Presets'] = require('./views/atoms/create_room/Presets');
 skin['atoms.create_room.RoomAlias'] = require('./views/atoms/create_room/RoomAlias');
