+++ src/skins/vector/views/molecules/ChangePassword.js
@@ -62,7 +62,7 @@ module.exports = React.createClass({
                     </div>
                 );
             case this.Phases.Uploading:
-                var Loader = sdk.getComponent("atoms.Spinner");
+                var Loader = sdk.getComponent("elements.Spinner");
                 return (
                     <div className="mx_Dialog_content">
                         <Loader />
